requirements.txt
quart>=0.18.4
aiohttp>=3.8.1
python-dotenv>=0.19.2
openai>=0.27.0
google-cloud-language>=2.4.1
supabase>=0.7.1
praw>=7.6.0
tweepy>=4.12.0
beautifulsoup4>=4.9.3
hypercorn>=0.14.3
asyncio>=3.4.3
stability-sdk>=0.8.0
pexels>=0.1.0
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
beautifulsoup4>=4.9.3
requests>=2.31.0
schedule>=1.2.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
